# company/specification/src/main/resources/api.yaml
openapi: 3.0.2
info:
  title: "E-Stock listing company API"
  description: "Register company in stock market"
  version: "0.0.1-SNAPSHOT"
servers:
  - url: "https://localhost:8081"
    description: "Development server"
paths:
  /api/v1.0/login:
    post:
      summary: Login to company application
      operationId: getAccessToken
      tags:
        - Login
      requestBody:
        description: Credentials payload
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Created token succcessfully
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Token'
  /api/v1.0/market/company/register:
    post:
      summary: Register a new company
      operationId: createCompany
      tags:
        - Company
      requestBody:
        description: Company payload
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/Company'
        required: true
      responses:
        '200':
          description: Company succesfully registered
          content:
            text/plain:
              schema:
                type: string
  /api/v1.0/market/company/info/{companyCode}:
    get:
      summary: Get company information by its code
      operationId: getCompanyByCode
      tags:
        - Company
      parameters:
        - name: companyCode
          in: path
          description: 'To get company information need company code'
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: 'Company details reterived successfully'
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Company'
  /api/v1.0/market/company/getall:
    get:
      summary: Get companies
      operationId: getCompanies
      tags:
        - Company
      responses:
        '200':
          description: 'Company details reterived successfully'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/Company'
  /api/v1.0/market/company/delete/{companyCode}:
    delete:
      summary: Delete company information by its code
      operationId: deleteCompanyByCode
      tags:
        - Company
      parameters:
        - name: companyCode
          in: path
          description: 'To get company information need company code'
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: 'Company deleted successfully'                                
components:
  schemas:
    Credentials:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    Token:
      type: object
      properties:
        token:
          type: string
    Company:
      type: object
      required:
        - code
        - name
        - ceo
        - turnover
        - website
        - stockExchange
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        ceo:
          type: string
        turnover:
          type: integer
          minimum: 1000
          example: 1000
        website:
          type: string
        stockExchange:
          type: string
        stockPrices:
          type: array
          items:
            type: object
            $ref: '#/components/schemas/StockPrice'
    StockPrice:
      type: object
      required:
        - stockPrice
        - startDate
        - endDate
      properties:
        id:
          type: integer
        stockPrice:
          type: float
        startDate:
          type: string
          format: date
          example: "2021-01-30"  
        endDate:
          type: string
          format: date
          example: "2021-01-30" 
         